- Added directories for each page - Moved router to App - Renamed router to TempNav (to be replaced by Navigation component)
