+++ src/ApiKey.env
@@ -1,0 +1,1 @@
+﻿export apiKey="COgbEuXTLZJ5nJkPgiJ68Uz7RaAcwp6z0GCrcoNj"
