+++ src/Pages/CuriosityLandingPage/CuriosityLandingPage.tsx
